requirements.txt
discord.py[voice]
substring
requests
FFmpeg  
youtube_dl
ffprobe 
validators 
youtube-search 
aiohttp==3.6.0
